Allow go_genrule to work from external dependencies.
The previous code assumed that `$(pwd)` was the root of the workspace that
uses `go_genrule`. This isn't true when using Bazel's external
dependencies, which place the code under `external/${DEP_NAME}/`.

This change is a bit hacky. Ideally the `go_genrule` implementation
wouldn't make assumptions about the code layout and would use
`$(location)` to find paths, but that would be a much larger change.
Special-casing build paths starting with `external/` was easier and
non-invasive.